+++ gbdxtools/ipe/error.py
@@ -16,3 +16,10 @@ class Unauthorized(Exception):
 
 class Forbidden(Exception):
     pass
+
+
+class MaxTries(Exception):
+    pass
+
+class UnsupportedImageType(Exception):
+    pass