+++ tests/unit/test_dem_image.py
@@ -1,0 +1,66 @@
+'''
+Authors: Donnie Marino, Kostas Stamatiou
+Contact: dmarino@digitalglobe.com
+
+Unit tests for the gbdxtools.Idaho class
+'''
+
+from gbdxtools import Interface
+from gbdxtools import DemImage
+from auth_mock import get_mock_gbdx_session
+import vcr
+from os.path import join, isfile, dirname, realpath
+import tempfile
+import unittest
+import rasterio
+import dask.array as da
+
+try:
+    from urlparse import urlparse
+except:
+    from urllib.parse import urlparse
+
+def force(r1, r2):
+    return True
+
+my_vcr = vcr.VCR()
+my_vcr.register_matcher('force', force)
+my_vcr.match_on = ['force']
+
+# How to use the mock_gbdx_session and vcr to create unit tests:
+# 1. Add a new test that is dependent upon actually hitting GBDX APIs.
+# 2. Decorate the test with @vcr appropriately
+# 3. Replace "dummytoken" with a real gbdx token
+# 4. Run the tests (existing test shouldn't be affected by use of a real token).  This will record a "cassette".
+# 5. Replace the real gbdx token with "dummytoken" again
+# 6. Edit the cassette to remove any possibly sensitive information (s3 creds for example)
+
+
+class IpeImageTest(unittest.TestCase):
+
+    _temp_path = None
+
+    @classmethod
+    def setUpClass(cls):
+        mock_gbdx_session = get_mock_gbdx_session(token='dymmytoken')
+        cls.gbdx = Interface(gbdx_connection=mock_gbdx_session)
+        #cls.gbdx = Interface()
+        cls._temp_path = tempfile.mkdtemp()
+        print("Created: {}".format(cls._temp_path))
+
+    @my_vcr.use_cassette('tests/unit/cassettes/test_dem_image.yaml', filter_headers=['authorization'])
+    def test_dem_image(self):
+        bbox = [-109.72, 43.19, -109.49, 43.34]
+        img = self.gbdx.dem_image(bbox=bbox)
+        self.assertTrue(isinstance(img, DemImage))
+        assert img.shape == (1, 554, 851)
+        assert img.proj == 'EPSG:4326'
+
+    @my_vcr.use_cassette('tests/unit/cassettes/test_dem_image_proj.yaml', filter_headers=['authorization'])
+    def test_dem_image_proj(self):
+        bbox = [-109.72, 43.19, -109.49, 43.34]
+        img = self.gbdx.dem_image(bbox=bbox, proj="EPSG:3857")
+        self.assertTrue(isinstance(img, DemImage))
+        assert img.proj == 'EPSG:3857'
+
+