+++ changelog.md
@@ -1,3 +1,9 @@
+0.12.0
+-------
+* upgrading to new dask version support (dask >= 0.15.1)
+* using the __new__ pattern for image classes
+* fixes for creating geotiffs from images  
+
 0.11.14
 -------
 * orthorectify parameters bugfix